perf(data-room): streaming SSR — server-prefetch init state, hand to client as prop
Before this PR, /data-room was a single 4848-line client component
that rendered <DataRoomBootShell /> on mount and then fired
`getDataRoomInitState(...)` from a useEffect to load everything.
On a cold lambda that fetch took 2.5-4s (PR-38 measurements) —
pure iad1↔ap-south-1 RTT + connection-pool warmup, not Postgres
execution time. Users stared at the spinner for that entire window.

PR-44 converts the route into an RSC wrapper that calls
getDataRoomInitState during server-side render and passes the result
down as a serialized `initialInitResult` prop. The server already
paid the lambda-init + Prisma-pool-warmup tax to render the layout
(PR-33); folding the init fetch into the same critical path overlaps
it with React rendering instead of chaining it after hydration.

Structural change:
  - app/data-room/page.tsx → renamed to app/data-room/DataRoomClient.tsx
    (the full 4848-line client component, verbatim contents)
  - New app/data-room/page.tsx is an RSC wrapper that:
      • awaits Next 15's `searchParams` Promise
      • forwards company_id into getDataRoomInitState
      • catches any prefetch error and falls back to null
      • renders <DataRoomClient initialInitResult={...} />

Surgical client-side change:
  - DataRoomPageInner now accepts `initialInitResult` prop
  - The init useEffect's `await getDataRoomInitState(...)` is gated:
    if the prop is present AND it matches the URL's company_id
    (or the URL omits company_id), use the prop directly; otherwise
    fall through to the original client fetch path.

Failure mode preserved: when the server prefetch throws (cookie
missing, transient network), the RSC passes `null` and the client's
useEffect re-fetches exactly as before — no regression versus the
pre-PR-44 behavior.
